prop-inspector: Don't notify the control changes while reloading
When the property inspector reloads the property into the edit control
it now sets a flag so that when the property changes in the control it
will temporarily ignore it. Otherwise it will end up notifying the
property change to rig which will put an extra entry in the undo
journal.