Simplify the undo journal
Previously the undo journal had two lists, one for the undo entries
and one for the redo entries. It also had a pointer within the undo
list for the current position. As far as I could tell, this position
marks the start of the redo entries within the undo list. It seems
kind of confusing to have the redo entries effectively stored in two
places so this patch removes the position so that there are just the
two lists. Undoing always means popping an operation off the end of
the undo list and redoing always means popping off the end of the redo
list. There are now also some comments to describe how the lists work.
The GQueues have been replaced with RutLists and the operations now
have an embedded list node.