Fix the signature for the rig_node_free function
The rig_node_free function was originally just used as a callback for
iterating the GQueue in a RigPath so it took the node type wrapped in
a gpointer. This function was moved over to rig-node.c but it kept the
same weird signature. This patch changes it to take a proper
RigPropertyType argument and adds a wrapper callback for the list
iterator to convert the arguments.