undo-journal: Log constant property changes and timeline changes
Instead of just logging changes to an entity's current property value,
the undo journal now tries to log changes to the transition. This
means it has separate operations to update the constant value and
add/remove/modify the property's path. All of these operations are
triggered by calling rig_undo_journal_set_property_and_log. This takes
a property and a boxed value and will make the right kind of undo
entry depending on whether the property is currently animated or not.
It will also make the necessary change to the transition.

Rig now uses this new function when the user finishes dragging an
entity or when updating a property from the inspector. This has the
side effect that the inspector will now correctly modify the timeline
instead of just the property's current value.