Remove the existing timeline view
This removes the code to to display and manipulate the timeline at the
bottom of the editor window. The code to draw the path no longer works
anyway and the space will be for a more complete path editor. The
timeline view is now just added to one of the RutSplitViews and is
drawn with the normal camera.