Add a RigTransitionView object for the timeline view
This adds a new control called RigTransitionView which displays of all
the animated properties for a transition. The view initially queries
the given transition for all animated properties and then registers a
callback to get notification for when any property toggles its
animated state. The list of properties is updated as the properties'
animated state is toggled.

The properties are grouped according to which object they belong to.
The name for each object and each property is listed in a column on
the left. There code is written so that it can layout this list in
columns with the expectation that we will later added more controls
per property.

To the right of the list is a horizontal section for each property
which contains diamond dots representing the key frames along the path
for that property. The dots are renderered as point sprites from a
vertex buffer. The idea is that this vertex buffer gets updated
whenever a property or a path node is added or removed.

The dots are rendered in a clipped region next to each property. The
scale for the dots is set up such that 0 is the left of the display
and 1 is the right.

The timeline progress is shown in the transition view with a vertical
semi-transparent red bar. The position can be changed by clicking and
dragging the mouse.