Remove the list of lights
RigData had a list of lights but it was only used in a block of code
that was commented out and passed to a function with no definition.
The loader code was trying to maintain this list of lights as a
separate list from the list of entities. However, the code to prepend
to the list of lights looked quite broken and instead it was
prepending to the list of entites but storing the head pointer in the
lights list. Presumably this would mean the lights entities would just
effectively disappear from the list of entities unless you happen to
add two lights in succession and walk the list of entities backwards.

Instead of making the list of lights work properly this just removes
it because it appears to be dead code anyway.