rut-graphable: Fix calling the child_removed method
RigGraphable has a method to notify when a child is added. However it
was only being called if the child was being moved to another parent,
not if it was just being removed. Also it looks like when the child is
moved it wasn't actually being removed correctly but the method just
gets called directly. This changes it so that the method is called
from rut_graphable_remove_child and rut_graphable_add_child now calls
the remove child function if the child already has a parent.