rut-text-buffer: Make the text property setter take only one argument
This changes the rut_text_buffer_set_text function to only take a
single argument which is the string containing the text. This also
adds a second function called rut_text_buffer_set_text_with_length
that takes a length parameter like the old text setter. This is
necessary because the setter is used as the function in the
RutPropertySpec and those functions only expect one argument.