Adds RutMesh utility api
This adds some utility apis for dealing with arbitrary meshes optimized
for memory usage and layouts suitable for GPUs but intended for CPU
access to the data. A RutMesh is comprised of N RutAttributes that
reference RutBuffers, very similar to a CoglPrimitive except the data is
directly accessible on the CPU. A CoglPrimitive can easily be derived
from a mesh. Utilities for iterating all vertices of a mesh or iterating
all the triangles of a mesh are provided by this api.