Associate tags with assets
We want to be able to search assets based on tags so this adds api to
associate assets with tags. Note the tags here are referred to as
"directory tags" because we are currently only looking at associating
tags derived from an assets directory location with the asset and we
anticipate that these tags should always be maintained separately from
user defined tags so that user defined tags can be saved to xml. If we
were to save location derived tags to xml then then we'd get in a
muddle if an asset was ever moved.