ui-viewport: Set the size of sizable children when size changes
When the size changes on a RutUiViewport it now sets the size of all
the direct children to be their preferred size. If the viewport is not
pannable in a direction then it will instead set the child to the size
of the viewport. This gives the children the opportunity to know how
big they should render. This will be used in the RigTransitionView.