mesh: Fix iterating the triangles
To iterate the triangles, rut_mesh_foreach_triangle sets the initial
triangle based on the first three vertices and then jumps to the next
triangle based on the vertex mode of the mesh. However it was not
calling the callback in between the first triangle and the second so
it would be ignored. Also because ‘i’ variable was in a for loop it
would increment it four times instead of three when the vertex mode is
triangles. Presumably this would make it skip a percentage of the
triangles at the end and also to use the wrong indices.

This patches changes it to use an infinite loop which always calls
the callback at the start. That way it should always catch the first
triangle. The loop terminator is only checked after the callback is
invoked. The ‘i’ variable is now only incremented in the handlers for
each vertex mode so that they can increment it by the right number of
vertices.

This was causing a bug where picking would only work if you click on the
right-hand side of diamond.