Initialise the framebuffer matrix before picking
The picking code uses the framebuffer matrix to track the object's
transforms while traversing the scene graph. However it wasn't
initialising it to anything so it would just have whatever camera
happened to last flush the matrix. This works out because the timeline
view happens to be last and happens to have the identity matrix.