mesh-ply: fix matching and setup of attributes
This fixes a bug with matching and setting up attribute descriptions
in cases where some attributes are found to be missing followed by other
attributes that have been found. We were assuming we could index the
original array of given attributes to find property names at a point
where we know that some of the original attributes are missing. We now
setup the loader_properties earlier so they can be referenced instead.