main: Fix muddling the timeline elapsed value and the progress
When the timeline progress changes, Rig was using the timeline elapsed
value to set the transition progress. It seems to make more sense to
use the progress value.

This makes it so that the progress value on the transition is
effectively always normalised in the range 0→1. This also means that
any new node in the transition will be stored with a normalised t
value and saved in the XML that way. I think it makes sense to keep
the animations all normalised and then you can have a separate
property on the timeline to specify the overall length of the
animation. That way the animation length can be adjusted without
having to move all of the nodes. Also it makes more sense if the
transition is not being controlled by time but for example by the
relative position of a swipe.