transition-view: Allow selecting nodes
The transition view now maintains a linked list of selected path
nodes. The nodes can be selected by clicking on them. If shift is held
down then multiple nodes can be selected. The selected nodes are drawn
white instead of red. The node colour is now stored in the attribute
buffer for the dots.

Currently selected a node is totally pointless because nothing can be
done with them.