Adds scroll bar support to ui viewports
This adds basic support for displaying scrollbars on viewports when the
document being scrolled is larger than the viewport itself. This splits
the RutScrollBar widget out from rut-shell.c and adds some missing
scroll-bar capabilities to allow viewports to manage x any y scroll bars
according to the current document size/position and viewport size. The
viewport will only show scroll-bars for pannable axis.

Note: Currently the document size has to be specified explicitly by
whatever code is responsible for populating the viewport with children,
but later we may want to instead map the size to the preferred size of a
child implementing the sizeable interface.