Add a RutInputable interface
This interface represents an object that wants to receive input. There
are no methods on the interface but instead there is just some
property data pointing to the input region. RutInputRegion itself
implements this and sets the input region pointer to point to itself.

The idea is that this way the input regions can be more tied to the
nodes of widgets. This will make it easier to bubble unhandled events
up the hierarchy.