Remove the transform from input regions
The RutInputRegion had an optional transform. However this was not
used anywhere and it looked like it would only work when the input
region does not have a parent because the picking code only checks for
a transform after checking for a parent. It would probably make more
sense to just add the input region to a RutTransform parent if this
functionality was needed.