Remove the global shell input regions
Previously the RutShell had a global list of input regions that was
only used by RutTool. This would check the input region with the
identity transform. The picking code had a kind of hack just for
circle regions so that it could accept NULL cameras. The RutTool can
get the same affect by setting HUD mode on its input region. Any
camera can be used because it is effectively ignored in HUD mode. The
code that adds the input region already has a pointer to a camera from
the camera component so it might as well use that.