Replace rut_shell_pick_input_region with rut_shell_pick_inputable
The main picking function has been replaced with an equivalent
function that can pick on any inputable object. As an input region
also implements the inputable interface code can just continue passing
input regions to the function. When the inputable is also graphable
the transform is taken from the inputable object not from the input
region itself.