Bubble events in the scenegraph instead of sending to all
The code to handle events for cameras is now split into two phases.
Firstly the input regions are checked and only if none of the input
regions for any of the cameras claim the event then the scenegraph of
each camera will be checked. Previously these two phases were
interleaved.

The phase which checks the scenegraphs has been modified so that
instead of sending the event to everything in the graph until an input
region claims the event, it now selects a single node to send the
event to using a depth first search. The input regions are checked in
the post_children callback of the search so that the deepest node will
be found except that it doesn't worry about trying to find a deeper
node that might be overlapping this node in a sibling branch. If the
selected node does not handle the event then it will also be bubbled
up to its ancestors until one of them handles it.

Keyboard events also use this bubbling mechanism. Additionally if
nothing has the keyboard focus then the node will be selected based on
the input region that intersects with the last known position of the
mouse. This effectively implements a kind of sloppy focus that has a
lower precedence than the normal keyboard focus. Places that were
previously grabbing the keyboard focus on motion events have been
removed because they can now rely on the default sloppy focus instead.

The RutUIViewport now implements the inputable interface instead of
adding its input region as a child of the viewport. This allows it to
handle the middle mouse button and the page up and down keys even if
there is a child widget with keyboard focus or an input region because
the unhandled events will bubble up the viewport.