Add a rig-types.h to workaround circular header dependencies
The RigData struct has pointers to many other different structs. Most
of these structs have a constructor that also takes a pointer back to
the RigData struct. This creates circular header depedencies for
example between rig-undo-journal.h and rig-data.h. Previously the
circle was broken for that example by declaring the typedef for
RigUndoJournal in rig-data.h. This seems a bit messy and it makes it
difficult to have a header that needs to refer to RigUndoJournal but
not RigData. The usual solution to this problem aside from making a
better layering abstraction is to just stuff all of the typedefs in a
separate header called foo-types.h. This patch does that by adding
rig-types.h which currently only has a typedef for RigData. That way
any types that need to refer to a RigData pointer in their constructor
can just include rig-types.h instead of rig-data.h. rig-data.h can
then just include the appropriate headers to get typedefs for the
pointers it needs.