Make some key events properly global
The keys for undo, redo and save are now handled in the
shell_input_handler callback instead of the editor input callback so
that they will work regardless of which object has the key focus.
Previously they only worked when the mouse was over the editor region.
That didn't make much sense considering the editor region is not the
only way to make changes that could be undone or saved.