path: Reset the position cache when its node is deleted
RigPath has a ‘pos’ pointer which seems to be used so that searching
for nodes around a particular time can be sped up by assuming that
subsequent searches will be at a nearby position. This cache wasn't
getting cleared if the cached node is removed which would later make
it crash.