transition-view: Fix updating has_selected_nodes cache on delete
When a node is deleted it was previously always setting the
has_selected_nodes cache for a property to FALSE even if there were
other selected nodes for that property. This patch fixes that by
removing the break statement so that it will always walk to the entire
list of selected nodes to correctly update the has_selected_nodes
value.