stack: Fix tracking the children
rut_stack_append_child was adding the child to the graph and then
additionally adding the child to its own list of children. However,
the added callback for the graphable interface would also add the
child to the list so it would end up there twice. This patch just
removes rut_stack_append_child because nothing was using it and
everything might as well just use the graphable interface directly to
add children.

The destructor for the stack was unreffing all of the children. This
shouldn't be necessary because the graphable state owns the reference
for the children and they will be unreffed in rut_graphable_destroy.

This patch also converts the list of children to be a RutList. The
main incentive for doing this is so that I can later add extra
per-child data to the list node.