split-view: Fix the ref counting for the two children
When one of the two children is set on the split view it takes a
reference. However it was not unreffing the previous child if a new
child is set. This fixes it to unref the child and also makes the free
function explicitly set the children to NULL so that it can share the
unreffing code.