Remove various junk from rut.c
There were a lot of old comments and an unused #if 0'd text widget
implementation in rut.c that this patch simply deletes. There are a few
comments at the beginning of the file that should probably be moved
somewhere else more useful or deleted later if not so these have been
kept for now.

A whole load of comments related to the paint-batchable stuff I had
looked at early on in the project have been deleted since we aren't
using any of this stuff in Rig currently. This is noted just in case we
do look at the paint batchable ideas again later and may want to recover
the comments.