shell: Add a callback mechanism to be invoked just before painting
This adds a mechanism to register a callback on the RutShell that will
be invoked just before the shell is painted. The callback is
associated with a graphable object so that the callbacks can be
invoked in order of increasing depth of the objects within their
scenegraph. Only one callback per graphable is allowed.

The intended usage of this mechanism is so that objects can delay
layouting their children until just before painting. Whenever a
property that changes a widget's layouting is altered, the widget is
expected to register the callback. In the implementation of the
callback it will actually layout its children according to its current
size. Usually changing a widget's size will affect the layout so it is
important that the callbacks are invoked in increasing order of depth
so that the children will have the correct size while they are
allocating.

It is safe and expected that an object will queue a pre-paint callback
during the invocation of another pre-paint callback.

The sorting of the list of callbacks is delayed until the queue is
just about to be flushed so that it doesn't matter if an object moves
in the hierarchy before the scene is actually painted. However if a callback
is added while the list is being flushed then it will be inserted in
the right place according to its depth. This assumes that widgets
won't change their parents during the pre-paint callback.

If a child is removed from the hierarchy while it still has a callback
registered then nothing is done to try and remove the callback. I
think this makes sense because it won't do any harm to call it even if
the graphable is not really in the scene and it may be that the
object's parent doesn't affect its layouting so the callback will do
the right thing anyway.