Add a sizable method to register for preferred size notifications
This adds a method to the sizable interface to register for
notifications whenever the object's preferred size changes. The method
returns a RutClosure pointer like other callback registration
functions.

The implementation of this method is optional. If it is not
implemented then the wrapper function will create a dummy closure
object that will never be invoked so that the rest of the code does
not have to care if the object has an implementation or not.

Currently there are no implementations but eventually it might make
sense to make all of the widgets start using this.