viewport: Add properties to sync the doc size to child's pref size
This adds a property called ‘sync_widget’ which will cause it to
make the doc size track the preferred size of the given widget.

Originally I thought it would be simpler to make this property a
boolean and have it track the size of the first child. However this is
a bit fiddly because the children aren't actually children of the
viewport but instead they are children of the doc node. All of the
code using the viewport gets a pointer to the doc node and directly
adds children to it. The viewport can't get notification of a child
being added so it's difficult to register the preferred size callback.

Additionally there are two properties called ‘expand_x/y’ which
specify whether the child should be set to larger than its preferred
size if the viewport itself is larger.