Adds alpha mask asset support
This adds support for an alpha-mask texture asset type. Such assets
should be placed in an alpha-masks/ directory so they are automatically
tagged as alpha masks. Attaching an alpha mask to an entity lets you
mask out a portion of the entity according to what areas of the mask's
alpha channel are below the current threshold associated with the
entity's material.