viewport: Update the cache of whether the scroll bars are visible
When the scroll bars are repositioned, there is a check for whether
the scroll bars are already visible and if it is the same as the new
value then it won't add or remove them. However the cache wasn't being
updated so if the scroll bar ever becomes visible then it will never
be removed.