shell: Do an exhaustive search when looking for an event target
When the scroll bars are displayed in a viewport, they will be
positioned on top of the viewport's document which means the input
region for children of the viewport will likely intersect with the
input region of the scroll bar. Previously the children of the
viewport would get priority over the scroll bars because the search
stopped at the first input region that contains the point. This patch
changes it to do an exhaustive search and pick the last input region
that contains the point so that the scroll bars will be selected
instead. This should more accurately reflect the order that the
widgets are painted so it is more likely that it will pick the thing
painted on top.