rut-tool: Take a grab when the rotation tool is being dragged
The rotation tool has a circular input region in screen space which is
used to detect clicks on the circle. Previously this circle is given a
radius of 64 when the mouse button is up and 128 when it is down. This
meant that while dragging if you moved the mouse outside of the larger
circle then it would ignore the input. I think it makes more sense to
just accept input anywhere and this is at least what Blender does.

It looks like it also had a problem where when the mouse button is
released it would reset the center of the input region to the last
mouse event position. Presumably if you let go of the mouse at the
edge of the circle then it would end up in the wrong place.