tool: Notify changes to the rotation via a callback
Instead of directly altering the rotation property on the entity when
the rotation tool is manipulated, it now just reports the desired
changes through a callback mechanism. The callback includes
information to specify whether the mouse is being dragged or whether
the button has been released so that the callback knows whether to
commit the changes or not. Rig now uses this to either just update the
property or to add the entry in the undo journal. Adding to the undo
journal will also cause it to create a transition path node if the
rotation property is currently animated.