Remove the cogl_foo_init_from_foo copy constructors
The CoglQuaternion type is allocated by the caller and can not have
any additional resources so it should be obvious that you can just use
the regular assignment operator to copy them. None of the other types
except the commented out vec4 type have a copy constructor so we
might as well remove these too.

Reviewed-by: Robert Bragg <robert@linux.intel.com>