text: don't swallow clicks that ungrab focus
RutText grabs all input when selected but if you click outside the text
entry then the grab is release. This makes sure a click outside the text
entry doesn't get marked as handled so for example if the user goes to
grab a scrollbar while a text entry has focus then they don't need to
click twice, once to release the grab and again the grab the scrollbar.