entity: allow multiple pipelines to be cached
Each entity now has 2 slots for caching pipelines which the rendering
code can use as it likes. This reworks the Rig rendering code to also
cache the pipelines used for rendering the shadow maps with the
entities. Caching will be needed if the shadow map pipeline needs to be
tweaked per-entity which it will need to be to support casting shadows
according to the alpha mask of the entity.