rig: splits out entity rendering code
This splits out the code for rendering a scene of entities into
rig-renderer.c to help keep main.c and the rendering code itself
maintainable. Also at some point we want to be able to swap in
completely different renderers suited to different visual ui styles
so it makes sense to keep this code orthogonal.