Allow removing arbitrary grabs in a grab input callback
When the number slider is clicked it takes a grab on the input,
creates a text widget and then calls rut_text_grab_key_focus. The text
widget actually then takes a full grab rather than just grabbing the
key focus. When the escape key is pressed the number slider will
handle the key press in its grab handler which will cause it to
destroy the text widget and release its own grab. The text widget will
also release its grab on destruction which means two grabs are
released within a single grab callback. This was causing it to crash
because RutShell only had code to cope with removing the grab that is
currently being invoked.

This patch changes it so that the list iterator variable for the
callbacks is now attached to the RutShell rather than being a local
variable. That way the function to remove a grab can detect if the
grab being removed is the one that is going to be processed next and
in that case it can move the iterator variable along by one.

This patch also changes the grab list to be a RutList instead of a
GList for no particular reason.