Don't trash the alpha value during lighting/shadowing
This makes sure we don't modify the alpha value of a fragment during
lighting and shadowing, keeping the alpha of the primary texture
combined with any shape/alpha mask. Previously the specular lighting,
for example, was causing fragments that would have basically been
invisible due to their transparency becoming black and opaque in the
final rendered results.