undo-journal: Log property changes for any object, not just entities
All of the functions to modify properties in the undo journal
previously took an entity argument which was meant to specify the
entity that the property belongs to. However, the undo journal is also
used to modify properties of components so this didn't make a whole
lot of sense. It was still working anyway because the entity argument
was only used to take a reference to something in order to keep the
property alive. This patch removes the entity arguments from all of
the functions and instead the object to reference is taken from the
object pointer in the property. The internal entity member of the undo
entries has been renamed to ‘object’.