prop-inspector: Destroy the dummy property
The dummy property used to listen for changes to the widget was
leaking when the property inspector was destroyed.

This also fixes a problem where destroying some widgets causes
properties to be notified which where in turn causing Rig to handle
the change and update the widget. Destroying the property before
destroying the widgets breaks the link so the notifications will be
ignored.