entity: Return the empty string instead of NULL for the label
The getter for the label property was returning NULL unless something
had been set on it. This makes it difficult to set a property which
depends on the label property because it would have to handle the NULL
cases. This patch makes it return the empty string instead in that
case. The entity saving code now also checks the empty string when
deciding whether to save the label.