rut-entry: Don't draw the circles outside of the allocated size
The size of the text widget was being set to exactly the same size as
the entry widget. The entry widget would then draw the circles either
side of the text widget. If the entry widget is used in a layout which
places it at its preferred width then it would draw outside of its
allocated size and blat over neighbouring widgets. This patch changes
it so that the text is translated by half the allocated height and
shrunk so that it fits within the inner rectangle. The painted circles
are now painted within the allocated size.

Previously the painting code was drawing a complete circle at either
side of the entry. Half of each circle would then be overpainted with
the center rectangle. This optimises it a little bit to just draw
semi-circles at either side.