transition-view: Use the allocated height
The transition view now sets the input region to the size of the
allocated height rather than to the bottom of the last row. This fixes
updating the timeline progress when there are no properties.

The point size for the dots is no longer set if there are no rows
because this was causing it to set an invalid point size which was
generating GL errors.