Add a RutImage widget to draw a single CoglTexture
This adds a widget which can draw a given texture. The preferred size
of the widget is the size of the texture. If the widget is given any
other size then there is an enum property to determine what it should
do. It can either just scale the image to fit in the box, scale but
preserve the aspect ratio or repeat the image.