Add a grid behind the assets list to show transparency
There is now a RutImage attached at the back of the graph for the
assets list which contains a checkboard pattern to make transparent
parts of assets more obvious. The image is set to repeat and is drawn
at an extremely large size so that it will always fill the viewport.