Add a minimal journal for depth sorting
Instead of immediately drawing entities as we traverse the graph we now
log the entities into an array along with its modelview matrix.  When we
have fully traversed the scene we then sort the array so we can
correctly draw opaque objects front to back and transparent objects back
to front.