Combine assets from xml with those found automatically
We can't simply ignore the list of assets specified in a ui.xml file and
rely on finding assets automatically since the loaded ui will refer to
the assets found when loading and if those assets aren't known about
when re-saving later then we will end up with a corrupt file. We now
have some common code for inferring asset tags and loading an asset
which we use when loading assets from a ui.xml file and when we
automatically search for image assets. When automatically loading assets
we are careful not to load duplicate assets.

This patch changes how assets are serialized so asset elements no longer
have a "type" attribute and instead we rely on inferring the type.
Existing xml files can be updated with the following python snippet:

  import xml.etree.ElementTree as ET
  import sys

  for arg in sys.argv[1:]:
      tree = ET.parse(arg)
          for asset in tree.findall(".//asset"):
              del asset.attrib["type"]

          tree.write(arg)