Fix numerous compiler warning
There are still quite a lot of compiler warnings to fix, but this patch
makes a good dent. A lot of the warnings are about prototype missmatches
when initializing interface vtables where the interface is defined in
terms of a RutObject first argument but the implementations use a more
specific type for the first argument. Since we'd like to have the
compiler warn us about more legitimate prototype mismatches we don't
want to resort to using void * pointers in vtables and so the compromise
is that implementations should use RutObject *self arguments as their
first argument and simply cast internally to the more specific type.