Allow loading the scene light from xml
When loading a ui xml file if we find a light entity then we store that
in data->light. When continuing to initialize Rig we only create a
default light if one hasn't already been loaded. Since we don't
currently support loading/saving of camera state we always add a camera
component to the light during initialization.